klay.layered: - Fixed two compilation problems with the current Eclipse compiler.

+++ plugins/de.cau.cs.kieler.klay.layered/src/de/cau/cs/kieler/klay/layered/p2layers/HeuristicGeneralizedLayerer.java
@@ -102,8 +102,8 @@ public class HeuristicGeneralizedLayerer implements ILayoutPhase {
         
         // initialize some variables that can be configured using layout options
         random = layeredGraph.getProperty(InternalProperties.RANDOM);
-        wLen = layeredGraph.getProperty(Properties.EDGE_LENGTH_WEIGHT);
-        wRev = layeredGraph.getProperty(Properties.EDGE_REVERSAL_WEIGHT);
+        wLen = layeredGraph.getProperty(Properties.EDGE_LENGTH_WEIGHT).doubleValue();
+        wRev = layeredGraph.getProperty(Properties.EDGE_REVERSAL_WEIGHT).doubleValue();
         
    
         // ---------------------
